requirements.txt
tensorflow
pandas
scikit-learn
dash
dash-bootstrap-components
numpy
Pillow
